requirements.txt
fastapi
uvicorn
requests
python-dotenv
Jinja2
databases
